Flytter logikk fra api til serviceklassen. Forbedrer logging og fjerner trace logging som uansett ikke synes.
